requirements.txt
# Core dependencies
pip>=23.0.1
setuptools>=65.0.0
wheel>=0.40.0

# LlamaIndex and OpenAI
llama-index==0.10.27
llama-index-llms-openai==0.1.15
llama-index-embeddings-openai==0.1.7
openai>=1.1.0
python-dotenv==1.0.0

# PDF and Image processing
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
pypdf==4.0.1
PyMuPDF==1.23.8

# UI
streamlit==1.31.0
